+++ Data Warehouse (SQL)/2_DWHedwSQL.sql
@@ -125,8 +125,8 @@ CREATE TABLE [edw].[Fact_Measurements] (
  G_ID INT NOT NULL,
  D_ID INT NOT NULL,
  T_ID INT NOT NULL,
- Temperature DECIMAL(7,2),
- Humidity DECIMAL(7,2),
+ Temperature REAL,
+ Humidity FLOAT,
  CarbonDioxide INT
 );
 
@@ -144,7 +144,7 @@ CREATE TABLE [edw].[Fact_MoisturePots] (
  G_ID INT NOT NULL,
  D_ID INT NOT NULL,
  T_ID INT NOT NULL,
- Moisture DECIMAL(7,2)
+ Moisture FLOAT
 );
 
 ALTER TABLE [edw].[Fact_MoisturePots] ADD CONSTRAINT PK_Fact_MoisturePots PRIMARY KEY (MP_ID, P_ID, G_ID, D_ID, T_ID);
